requirements.txt
tweepy
python-telegram-bot
python-dotenv
